+++ server/src/main/java/dev/food/fast/server/auth/repository/UserRepository.java
@@ -2,13 +2,10 @@ package dev.food.fast.server.auth.repository;
 
 import java.util.Optional;
 
-import dev.food.fast.server.auth.dto.response.UserResponse;
 import dev.food.fast.server.auth.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
   Optional<User> findByEmail(String email);
