+++ server/src/main/java/dev/food/fast/server/general/repository/BasicOrderGuestRepository.java
@@ -3,5 +3,5 @@ package dev.food.fast.server.general.repository;
 import dev.food.fast.server.general.models.order.BasicOrderGuest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BasicOrderGuestRepository extends JpaRepository<BasicOrderGuest, Integer> {
+public interface BasicOrderGuestRepository extends JpaRepository<BasicOrderGuest, Long> {
 }
