+++ server/src/main/java/dev/food/fast/server/general/repository/ProductStatusLikesRepository.java
@@ -9,9 +9,9 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductStatusLikesRepository extends JpaRepository<ProductStatusLikes, Integer> {
-    Optional<ProductStatusLikes> findByProduct_IdAndUser_Id(Integer productId, Integer userId);
+public interface ProductStatusLikesRepository extends JpaRepository<ProductStatusLikes, Long> {
+    Optional<ProductStatusLikes> findByProduct_IdAndUser_Id(Long productId, Long userId);
 
     @Query("SELECT new dev.food.fast.server.general.dto.response.ProductStatusLikesResponse(pl.product.id, pl.status, p.likes) FROM ProductStatusLikes pl JOIN pl.product p WHERE pl.user.id = :userId")
-    List<ProductStatusLikesResponse> findResponseStatusLikedByUser_Id(@Param("userId") Integer userId);
+    List<ProductStatusLikesResponse> findResponseStatusLikedByUser_Id(@Param("userId") Long userId);
 }
