+++ server/src/main/java/dev/food/fast/server/general/repository/SliderImageRepository.java
@@ -3,6 +3,6 @@ package dev.food.fast.server.general.repository;
 import dev.food.fast.server.general.models.SliderImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SliderImageRepository extends JpaRepository<SliderImage, Integer> {
+public interface SliderImageRepository extends JpaRepository<SliderImage, Long> {
 
 }
